implement host statistic
git-svn-id: https://harrison.its-technidata.de:444/svn/ingrid/ingrid-iplug-se/trunk@13474 d5cb13e6-57fc-0310-9ab5-ddaedd36d86d

+++ plugins/opensource/admin-crawl/src/java/org/apache/nutch/admin/crawl/statistic/Statistic.java
@@ -3,10 +3,10 @@ package org.apache.nutch.admin.crawl.statistic;
 public class Statistic {
 
   private String _host;
-  private Integer _crawldbCount;
-  private Integer _segmentCount;
+  private Long _crawldbCount;
+  private Long _segmentCount;
 
-  public Statistic(String host, Integer crawldbCount, Integer segmentCount) {
+  public Statistic(String host, Long crawldbCount, Long segmentCount) {
     _host = host;
     _crawldbCount = crawldbCount;
     _segmentCount = segmentCount;
@@ -20,19 +20,19 @@ public class Statistic {
     _host = host;
   }
 
-  public Integer getCrawldbCount() {
+  public Long getCrawldbCount() {
     return _crawldbCount;
   }
 
-  public void setCrawldbCount(Integer crawldbCount) {
+  public void setCrawldbCount(Long crawldbCount) {
     _crawldbCount = crawldbCount;
   }
 
-  public Integer getSegmentCount() {
+  public Long getSegmentCount() {
     return _segmentCount;
   }
 
-  public void setSegmentCount(Integer segmentCount) {
+  public void setSegmentCount(Long segmentCount) {
     _segmentCount = segmentCount;
   }
